Split content ops connector packet contracts
