+++ loopx/capabilities/content_ops/schemas.py
@@ -1,0 +1,40 @@
+"""Schema identifiers for content-ops packets and state surfaces."""
+
+CONTENT_OPS_SURFACE_SCHEMA_VERSION = "content_ops_surface_v0"
+CONTENT_OPS_SURFACE_PROJECTION_SCHEMA_VERSION = "content_ops_surface_projection_v0"
+CONTENT_OPS_PREVIEW_PACKET_SCHEMA_VERSION = "content_ops_preview_packet_v0"
+CONTENT_OPS_PUBLIC_HANDLE_OBSERVATION_PACKET_SCHEMA_VERSION = (
+    "content_ops_public_handle_observation_packet_v0"
+)
+CONTENT_OPS_PUBLIC_HANDLE_OBSERVATION_SCHEMA_VERSION = (
+    "content_ops_public_handle_observation_v0"
+)
+CONTENT_OPS_PRIVATE_CONNECTOR_GATE_PACKET_SCHEMA_VERSION = (
+    "content_ops_private_connector_gate_packet_v0"
+)
+CONTENT_OPS_PRIVATE_CONNECTOR_OWNER_GATE_SCHEMA_VERSION = (
+    "content_ops_private_connector_owner_gate_v0"
+)
+CONTENT_OPS_CONNECTOR_RUNTIME_POLICY_SCHEMA_VERSION = (
+    "content_ops_connector_runtime_policy_v0"
+)
+CONTENT_OPS_PACKET_AGGREGATION_SCHEMA_VERSION = "content_ops_packet_aggregation_v0"
+CONTENT_OPS_CHATVIEW_CONNECTOR_REPORT_SCHEMA_VERSION = (
+    "content_ops_chatview_connector_report_v0"
+)
+CONTENT_OPS_WALKTHROUGH_ARTIFACT_SCHEMA_VERSION = (
+    "content_ops_walkthrough_artifact_v0"
+)
+CONTENT_OPS_EXPLORATION_PLAN_PACKET_SCHEMA_VERSION = (
+    "content_ops_exploration_plan_packet_v0"
+)
+EXPLORATION_PLAN_SCHEMA_VERSION = "exploration_plan_v0"
+
+SOURCE_ITEM_SCHEMA_VERSION = "source_item_v0"
+ANGLE_CANDIDATE_SCHEMA_VERSION = "angle_candidate_v0"
+DRAFT_ITEM_SCHEMA_VERSION = "draft_item_v0"
+FEEDBACK_SIGNAL_SCHEMA_VERSION = "feedback_signal_v0"
+PUBLISH_GATE_SCHEMA_VERSION = "publish_gate_v0"
+MATERIAL_MEMORY_SCHEMA_VERSION = "material_memory_v0"
+CONNECTOR_TRIAL_SCHEMA_VERSION = "connector_trial_v0"
+CONTENT_OPS_VALIDATION_SCHEMA_VERSION = "content_ops_surface_validation_v0"
